Add `indexing_view` module for rendering and handling codebase indexing UI. Refactor `main.rs` to use new module, improving modularity and code organization.
